+++ dashboard/src/views/dashboard/index.vue
@@ -94,14 +94,24 @@ export default {
       })
     },
     handleDelete(index, row) {
-      deleteApp({ appName: row.appName })
-        .then(response => {
-          console.log(response)
-          this.fetchData()
-          this.$message.success('删除成功')
+      this.$confirm('此操作将永久删除, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      })
+        .then(() => {
+          deleteApp({ appName: row.appName })
+            .then(response => {
+              console.log(response)
+              this.fetchData()
+              this.$message.success('删除成功')
+            })
+            .catch(() => {
+              this.$message.error('删除失败')
+            })
         })
         .catch(() => {
-          this.$message.error('删除失败')
+          this.$message.info('已取消删除')
         })
     },
     submitForm(form) {
